refactor services add tests

+++ CodingDojo/server/src/main/java/com/codenjoy/dojo/services/PlayerService.java
@@ -30,7 +30,7 @@ public interface PlayerService extends Tickable {
     Player register(String id, String game, String room, String ip,String repositoryUrl);
     Player register(PlayerSave save);
 
-    void updateScore(String gitHubUsername, double score);
+    void updateScore(String gitHubUsername, long score);
 
     List<Player> getAll();
     List<Player> getAll(String game);
